Adds FeatureSource
